+++ lib/qops/cookbook/cookbook.rb
@@ -130,10 +130,14 @@ class Qops::Cookbook < Thor
   end
 
   def s3
-    @s3 ||= Aws::S3::Client.new(
-      region: 'us-east-1',
-      credentials: config.opsworks.config.credentials.credentials
-    )
+    @s3 ||= Aws::S3::Client.new(**aws_configs)
+  end
+
+  def aws_configs
+    aws_config = { region: 'us-east-1' }
+    # use the profile if found
+    options[:profile] ? aws_config[:profile] = options[:profile] : aws_config[:credentials] = config.opsworks.config.credentials.credentials
+    aws_config
   end
 
   def local_artifact_file