+++ lib/qops/environment.rb
@@ -31,20 +31,69 @@ module Qops
       end
     end
 
-    def initialize
-      %w[deploy_type region stack_id app_name].each do |v|
+    def initialize(profile: nil, force_config: false)
+      @_aws_config = { region: configuration.region }
+      @_aws_config[:profile] = profile unless profile.nil?
+      @_force_config = force_config
+      puts Rainbow("using aws profile #{profile}").bg(:black).green unless profile.nil?
+
+      %w[deploy_type region app_name].each do |v|
         fail "Please configure #{v} before continuing." unless option?(v)
       end
 
+      fail 'Please configure the layer_name if you are allowing qops to search aws stacks' unless option?('layer_name')
+
+      # if being forced to use config , then stack_id is a requirement
+      fail 'Please configure stack_id or stack_name before continuing' unless option?('stack_id') || option?('stack_name')
+
       begin
-        opsworks.config.credentials.credentials
+        opsworks.config.credentials.credentials unless profile
       rescue => e # rubocop:disable Lint/RescueWithoutErrorClass
         raise "There may be a problem with your aws credentials. Please correct with `aws configure`. Error: #{e}"
       end
     end
 
-    def application_id
-      configuration.application_id unless configuration.application_id.blank?
+    def stack(options = {})
+      return @_stack if @_stack
+      # find out if the config is using stack id or name
+      key = search_key(options)
+      value = options[key] || configuration.send(key)
+      # aws uses the term 'name' to reference a stack name
+      key = :name if key == :stack_name
+      @_stack = search_stack(key, value)
+    end
+
+    def stack_id(options = {})
+      return configuration.stack_id if @_force_config
+      stack(options).stack_id
+    end
+
+    def subnet(options = {})
+      return configuration.subnet if @_force_config
+      stack(options).default_subnet_id
+    end
+
+    def layers(_options = {})
+      opsworks.describe_layers(stack_id: stack_id).layers
+    end
+
+    def layer_id(options = {})
+      return configuration.layer_id if @_force_config
+      name = options[:layer_name] || configuration.layer_name
+      layers.find { |layer| layer.name.casecmp(name) }.layer_id
+    end
+
+    def chef_version(options = {})
+      stack(options).configuration_manager.version.to_f
+    end
+
+    def apps(_options = {})
+      opsworks.describe_apps(stack_id: stack_id).apps
+    end
+
+    def application_id(options = {})
+      return configuration.application_id if @_force_config
+      apps(options).first.app_id
     end
 
     def deploy_type
@@ -67,8 +116,9 @@ module Qops
       configuration.autoscale_type || nil
     end
 
-    def opsworks_os
-      configuration.os || 'Ubuntu 14.04 LTS'
+    def opsworks_os(options = {})
+      return configuration.os if @_force_config
+      find_stack(options).default_os
     end
 
     # Default 1 days
@@ -85,11 +135,15 @@ module Qops
     end
 
     def opsworks
-      @_opsworks_client ||= Aws::OpsWorks::Client.new(region: configuration.region)
+      @_opsworks_client ||= Aws::OpsWorks::Client.new(**@_aws_config)
     end
 
     def ec2
-      @_ec2_client ||= Aws::EC2::Client.new(region: configuration.region)
+      @_ec2_client ||= Aws::EC2::Client.new(**@_aws_config)
+    end
+
+    def elb
+      @_elb_client ||= Aws::ElasticLoadBalancing::Client.new(region: 'us-east-1', profile: @_aws_config[:profile])
     end
 
     def cookbook_json
@@ -113,6 +167,32 @@ module Qops
     end
 
     private
+
+    def identity_from_config
+      configuration.stack_id ? :stack_id : :stack_name
+    end
+
+    def search_key(options = {})
+      if !options[:name].nil?
+        :name
+      elsif !options[:stack_id].nil?
+        :stack_id
+      else
+        id = identity_from_config
+        msg = Rainbow("Using opsworks.yml config #{id}: #{configuration.send(id)}")
+        puts(msg.bg(:black).green)
+        id
+      end
+    end
+
+    def search_stack(key, value)
+      stack = opsworks.describe_stacks.stacks.find { |s| s.send(key) == value }
+      unless stack
+        puts Rainbow("Could not find stack with #{key} = #{value}").bg(:black).red
+        exit(-1)
+      end
+      stack
+    end
 
     def method_missing(method_sym, *arguments, &block) # rubocop:disable Style/MethodMissing
       if configuration.respond_to?(method_sym)